requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
dnspython==2.7.0
email_validator==2.2.0
idna==3.10
pydantic==2.10.2
pydantic_core==2.27.1
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
